+++ paddle/phi/kernels/gpu/rnn_grad_kernel.cu.cc
@@ -221,7 +221,7 @@ void RnnGradKernel(const Context &dev_ctx,
       false,
       phi::errors::InvalidArgument("ROCm do not support SequenceLength yet."));
 #endif
-  std::vector<int> SequenceLength;
+  std::vector<int64_t> SequenceLength;
   if (has_seq_length) {
     phi::TensorToVector(*sequence_length, dev_ctx, &SequenceLength);
   }
